+++ src/main/java/com/petrolpark/util/BigItemStack.java
@@ -1,28 +1,28 @@
 package com.petrolpark.util;
 
+import java.util.ArrayList;
+import java.util.List;
+
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.level.ItemLike;
-
-import java.util.List;
-import java.util.ArrayList;
 
 public class BigItemStack {
 
     public static final BigItemStack EMPTY = new BigItemStack(ItemStack.EMPTY);
     
     protected final ItemStack stack;
-    protected final int count;
+    protected final long count;
 
     public BigItemStack(ItemStack stack) {
         this(stack, stack.getCount());
     };
 
-    public BigItemStack(ItemStack stack, int count) {
+    public BigItemStack(ItemStack stack, long count) {
         this.stack = stack.copyWithCount(1);
         this.count = count;
     };
 
-    public BigItemStack(ItemLike item, int count) {
+    public BigItemStack(ItemLike item, long count) {
         this.stack = new ItemStack(item);
         this.count = count;
     };
@@ -31,15 +31,15 @@ public class BigItemStack {
         return stack;
     };
 
-    public int getCount() {
+    public long getCount() {
         return count;
     };
 
     public List<ItemStack> getAsStacks() {
-        int leftOver = count % stack.getMaxStackSize();
+        long leftOver = count % stack.getMaxStackSize();
         List<ItemStack> stacks = new ArrayList<>(count / stack.getMaxStackSize() + leftOver == 0 ? 0 : 1);
-        for (int i = 0; i < count / stack.getMaxStackSize(); i++) stacks.add(stack.copyWithCount(stack.getMaxStackSize()));
-        if (leftOver > 0) stacks.add(stack.copyWithCount(leftOver));
+        for (long i = 0; i < count / stack.getMaxStackSize(); i++) stacks.add(stack.copyWithCount(stack.getMaxStackSize()));
+        if (leftOver > 0) stacks.add(stack.copyWithCount((int)leftOver));
         return stacks;
     };
 
@@ -47,7 +47,7 @@ public class BigItemStack {
         return stack.copyWithCount(count);
     };
 
-    public BigItemStack copyWithCount(int count) {
+    public BigItemStack copyWithCount(long count) {
         return new BigItemStack(stack, count);
     };
 };
